requirements.txt
Pillow
opencv-python
numpy
pyautogui
pyobjc-framework-Quartz
